src/flask/requirements.txt
Flask==1.1.2
Werkzeug==1.0.1
markupsafe==2.0.1

jinja2<3.1.0
itsdangerous==2.0.1
python-dotenv==0.20.0
PyMySQL==1.0.2 
pytest==7.1.2
selenium==4.1.5
docker==5.0.3
